refactor(api): split resetPassword from user service
